Make `tarantoolctl start` work withiout box.cfg in init script
`box.cfg` may not be called inside init script and tarantool
will not be daemonized (forcing `background = true` in hijacked `box.cfg`),
so we need to fork tarantool before running this script.

To find out the moment when tarantool finishes startup it waits for `READY=1`
message on NOTIFY_SOCKET which is emitted after init script is finished.