sql: fix default trim octet for binary strings
According to ANSI specification, if TRIM function accepts binary string
and trim octet is not specified, then it is implicitly set to X'00'.
Before this patch trim octet was set to ' ' both for string and binary
string arguments. In turn, ' ' is equal to X'20' in hex representation.
Hence, TRIM function cut wrong characters:

TRIM(X'004420') -> X‘0044'

This patch sets default trim octet to X'00' for binary string arguments.

Part of #4206