Export tarantool_package symbol
There is compile time option PACKAGE in cmake to define
current build distribution info. For community edition
is it "Tarantool" by default and "Tarantool Enterprise" for the
enterprise version.

It's displayed in console greeting and in `box.info().package`,
but, unfortunately, it can't be accessed before `box.cfg` neither
from Lua API, neither with FFI.

This patch exports `tarantool_package` symbol making it accessible with
FFI.