json: detect a new invalid json path case
JSON paths has no a strict standard, but definitely there is no
an implementation, allowing to omit '.' after [], if a next token
is a key. For example:

    [1]key

is invalid. It should be written like that:

    [1].key

Strangely, but we even had tests on the invalid case.

Closes #4419

(cherry picked from commit ef64ee51cc5820da63d66c8ba8baa62695549c2d)