staticbuild: fix RUN_TESTS condition
Before this patch RUN_TESTS condition in Dockerfile.staticbuild
was ignored and always was true

However adding of brackets solves only part of problem.
If RUN_TESTS is empty `sh -c` returns 1 and build fails.

However if we run tests we should fail build if tests
are not passed. Ternary logic was rewritten to fair if-else

This patch fixes it and allows build tarantool statically
without running tests