Make fibers joinable by default
"Joinable" flag at fiber is an implementation details
and should not bother end users.

This patch makes every new fiber joinable and registers
GC callback for it. GC'ed fiber can be either dead or
alive. As there is no more ref to the fiber from lua
it means that dead one can be recycled and alive one
can be marked as not joinable so it can be fully recycled
automatically later.

This makes set_joinable obsolete, however the patch
is completely backward-compatible.

@TarantoolBot document
Title: Remove set_joinable from docs or mark it obsolete
As of all fibers are joinable by default
set_joinable got useless.