Fix bootstrap.snap corrupted file.
Bootstrap.snap is created from a normal snapshot file, but with
erased VClock option in the header:

    SNAP
    0.13
    Version: 2.2.1-122-g1146bb78d
    Instance: 03d3836a-e608-421c-9f8d-ad9beefe7440
    VClock: {}

In a normal snapshot it is 'VClock: {1: ...}'. To erase the option
usually developers use 'vim'. But when a binary file is opened in
vim without any arguments, like this:

    vim bootstrap.snap

on close it will edit some parts of the file in unexpected ways,
depending on local vim settings.

To forbid any implicit changes binary mode should be used:

    vim -b bootstrap.snap

The patch regenerates bootstrap.snap and drops VClock using binary
mode vim.

Closes #4510