backtrace: fix out of bounds access on backtrace printing
snrpintf always null-terminates the passed string, and it also returns
the number of bytes that "would have been written if there was enough
space", so not only we don't have to null-terminate the string, but even
more so we shouldn't do it erroneously. The only case when a string
should be null-terminated manually is when the print cycle doesn't run
at all, so move the termination before the cycle.

Closes #4636

(cherry picked from commit 2066f297b88126fd600a195c9bc25e14b241b247)