clear terminal state on panic
The tarantool_free() call in the end of main() works all the time except
when we exit due to a panic. We need to clear terminal state in this
case also, so return to using atexit() to clear readline state.

Closes #4466

(cherry picked from commit c6d4f010633a9961233178666d5a3c056018a758)