sql: drop only generated sequence in DROP TABLE
It is possible to create a sequence manually, and give it to a newly
created index as a source of unique identifiers. Such sequences are not
owned by a space, and therefore shouldn't be deleted when the space is
dropped. They are not dropped when space:drop() in Lua is called, but
were dropped in SQL 'DROP TABLE' before this patch. Now Lua and SQL are
consistent in that case.